missing stuff.... still something missing...
git-svn-id: svn+ssh://iagpc151.iag.uni-stuttgart.de/usr/local/svnroot/Boltzplatz/branches/PICASSO@916 f8cfc4a6-4e3a-43c3-8588-810d72ca601e

+++ src/interpolation/basis.f90
@@ -955,7 +955,7 @@ INTEGER,INTENT(IN) :: N_in
 INTEGER(KIND=8)    :: FACTORIAL
 !-----------------------------------------------------------------------------------------------------------------------------------
 !local variables
-INTEGER            :: I
+INTEGER(KIND=8)    :: I
 !===================================================================================================================================
 IF(N_in.LT.0) CALL abort(__STAMP__,&
   'FACTORIAL of a negative integer number not allowed! ',999,REAL(N_in))
